Further tidy `tree_hash` crate.
